+++ DarkSideModernGUI/Views/Pages/DashboardPage.xaml
@@ -34,19 +34,10 @@
             Margin="0,0,0,0"
             HorizontalAlignment="Stretch"
             VerticalAlignment="Top"
-            DisplayMemberPath="GameDLLID"
-            SelectedValuePath="GameDLLID" />
+            DisplayMemberPath="procId"
+            SelectedValuePath="procId" />
 
 
-        <ui:Button
-            x:Name="btnInjectGameDLL"
-            Grid.Column="3"
-            Width="150"
-            Margin="33,4,0,0"
-            VerticalAlignment="Top"
-            Background="#FFFCA703"
-            Click="btnInjectGameDLL_Click"
-            Content="Inject Game.DLL" />
 
 
 
@@ -64,31 +55,7 @@
                 BorderThickness="1">
 
 
-                <StackPanel
-                    x:Name="Route_Load"
-                    Grid.Column="3"
-                    Margin="0,0,0,5">
 
-                    <Label
-                        Grid.Column="3"
-                        Width="180"
-                        Margin="10,10,0,0"
-                        HorizontalAlignment="Left"
-                        VerticalAlignment="Top"
-                        Content="Select Route"
-                        FontSize="14"
-                        FontWeight="Bold" />
-
-
-                    <ComboBox
-                        x:Name="cbxRouteLoad"
-                        Width="160"
-                        Margin="0,0,0,10" />
-
-
-
-
-                </StackPanel>
 
             </Border>
 
@@ -105,8 +72,61 @@
 
 
 
+        <StackPanel
+                    x:Name="Route_Load"
+                    Grid.Column="3"
+                    Margin="0,0,0,5">
+            <ui:Button
+            x:Name="btnInjectGameDLL"
+            Grid.Column="3"
+            Width="150"
+            Margin="33,4,0,0"
+            VerticalAlignment="Top"
+            Background="#FFFCA703"
+            Click="btnInjectGameDLL_Click"
+            Content="Inject Game.DLL" />
 
-        <DockPanel Grid.Column="3">
+            <ui:Button
+            x:Name="btnInjectAll"
+            Grid.Column="3"
+            Width="150"
+            Margin="33,10,0,0"
+            VerticalAlignment="Top"
+            Background="#FFFCA703"
+            Click="btnInjectAll_Click"
+            Content="Inject All" />
+
+            <ui:Button
+            x:Name="btnUnloadAll"
+            Grid.Column="3"
+            Width="150"
+            Margin="33,10,0,0"
+            VerticalAlignment="Top"
+            Background="#FFFCA703"
+            Click="btnUnloadAll_Click"
+            Content="Unload All" />
+
+            <Label
+                        Grid.Column="3"
+                        Width="180"
+                        Margin="10,10,0,0"
+                        HorizontalAlignment="Left"
+                        VerticalAlignment="Top"
+                        Content="Select Route"
+                        FontSize="14"
+                        FontWeight="Bold" />
+
+
+            <ComboBox
+                        x:Name="cbxRouteLoad"
+                        Width="160"
+                        Margin="0,0,0,10" />
+
+
+
+
+        </StackPanel>
+        <DockPanel Grid.Column="2" Grid.ColumnSpan="2" Margin="155,0,5,0">
             <Border
                 Width="200"
                 Height="230"
@@ -115,10 +135,12 @@
                 VerticalAlignment="Top"
                 BorderBrush="#FFFCA703"
                 BorderThickness="1">
+                
+                
                 <ui:Button
                     x:Name="btnLaunchBot"
                     Width="160"
-                    Height="160"
+                    Height="100"
                     Margin="0,30,0,0"
                     HorizontalAlignment="Center"
                     VerticalAlignment="Top"
